requirements.txt
alembic==1.5.8
aniso8601==9.0.1
bcrypt==3.2.0
bidict==0.21.2
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-DotEnv==0.1.2
Flask-Env==2.0.0
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
limits==1.5.1
Mako==1.1.4
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.12
Werkzeug==1.0.1
